move worker info
